WIP traversing the ast with a JSONVisitor
